composite and composer get their own files. stores update methods get their own class methods.
