+++ vivarium/plots/topology.py
@@ -11,7 +11,9 @@ from matplotlib.path import Path
 from matplotlib.figure import Figure
 import networkx as nx
 
-from vivarium.core.process import Process, Composer, Composite
+from vivarium.core.process import Process
+from vivarium.core.composer import Composer
+from vivarium.core.composite import Composite
 from vivarium.core.store import generate_state
 from vivarium.library.topology import normalize_path
 from vivarium.library.dict_utils import deep_merge
