+++ vivarium/processes/remove.py
@@ -10,8 +10,8 @@ from typing import Any, Dict
 from vivarium.core.experiment import pp
 from vivarium.core.process import (
     Deriver,
-    Composer,
 )
+from vivarium.core.composer import Composer
 from vivarium.core.composition import (
     simulate_composer,
     PROCESS_OUT_DIR,
