+++ vivarium/processes/swap_processes.py
@@ -9,8 +9,8 @@ from typing import Any, Dict
 
 from vivarium.core.process import (
     Deriver,
-    Composer,
 )
+from vivarium.core.composer import Composer
 from vivarium.core.composition import (
     simulate_composer,
     PROCESS_OUT_DIR,
